+++ R/org_sitrep.R
@@ -17,37 +17,50 @@ org_sitrep = function(org) {
     invisible(org)
   } else {
     admins = org_admins(org)
+    workflow_perm = org_workflow_permissions(org)
 
-    org = result(res)
+    res_org = result(res)
 
-    perm = org$default_repository_permission
-    if (perm != "none")
-      perm_warn = "members can currently view {.emph all} repos in this org."
+    repo_perm = res_org$default_repository_permission
+    if (repo_perm != "none")
+      repo_perm_warn = "members can currently view {.emph all} repos in this org."
     else
-      perm_warn = NULL
+      repo_perm_warn = NULL
 
+
+    if (workflow_perm != "write")
+      workflow_perm_warn = "this may prevent some GitHub actions from working correctly."
+    else
+      workflow_perm_warn = NULL
 
     #if (perm != "none")
     #  perm = cli::col_red(perm)
 
-    cli::cli_h1("{.strong {org$login} sitrep:}")
+    cli::cli_h1("{.strong {res_org$login} sitrep:}")
     cli::cli_ul()
     cli::cli_li(cli_kv("Admins", admins))
-    cli::cli_li(cli_kv("Members", org$collaborators))
-    cli::cli_li(cli_kv("Public repos", org$public_repos))
-    cli::cli_li(cli_kv("Private repos", org$total_private_repos))
+    cli::cli_li(cli_kv("Members", res_org$collaborators))
+    cli::cli_li(cli_kv("Public repos", res_org$public_repos))
+    cli::cli_li(cli_kv("Private repos", res_org$total_private_repos))
     cli::cli_li(
       cli_kv(
         "Default repository permission",
-        perm,
-        perm_warn
+        repo_perm,
+        repo_perm_warn
       )
     )
-    cli::cli_li(cli_kv("Members can create public repos", org$members_can_create_public_repositories))
-    cli::cli_li(cli_kv("Members can create private repos", org$members_can_create_private_repositories))
+    cli::cli_li(
+      cli_kv(
+        "Default workflow permissions",
+        workflow_perm,
+        workflow_perm_warn
+      )
+    )
+    cli::cli_li(cli_kv("Members can create public repos", res_org$members_can_create_public_repositories))
+    cli::cli_li(cli_kv("Members can create private repos", res_org$members_can_create_private_repositories))
     cli::cli_end()
 
-    invisible(org)
+    invisible(res_org)
   }
 }